Move codegenerator package to internal
This reduces our API surface
