+++ protoc-gen-grpc-gateway/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     srcs = ["main.go"],
     importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway",
     deps = [
-        "//codegenerator:go_default_library",
+        "//internal/codegenerator:go_default_library",
         "//internal/descriptor:go_default_library",
         "//protoc-gen-grpc-gateway/internal/gengateway:go_default_library",
         "@com_github_golang_glog//:go_default_library",
