+++ protoc-gen-swagger/main.go
@@ -9,7 +9,7 @@ import (
 	"github.com/golang/glog"
 	"github.com/golang/protobuf/proto"
 	plugin "github.com/golang/protobuf/protoc-gen-go/plugin"
-	"github.com/grpc-ecosystem/grpc-gateway/codegenerator"
+	"github.com/grpc-ecosystem/grpc-gateway/internal/codegenerator"
 	"github.com/grpc-ecosystem/grpc-gateway/internal/descriptor"
 	"github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger/internal/genswagger"
 )
